third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/eval-allowed-expected.txt

ALERT: PASS (1 of 2)
ALERT: PASS (2 of 2)
